stacklist: add DzlStackList animated stack list
This widget provides what we are calling a "Stack List". It
allows for diving down into a hierarchical data set similar to
a treeview, but without some of the difficulties of a treeview.

StackList does not indent as you dive down. You can only look
at a single parent in the hierarchy at a time. Generally, this
is what we want when we mistakenly use a treeview in a sidebar.

To be extra dazzling, we animate the row transition as we dive
to the next layer of the hierarchy.

This widget uses GListModel for content. Having a GListModel
of items which are also a GListModel is an ideal way to use
this widget. However, you probably want lazy initialization with
GAsyncIntiable for the descendants.