graphing: add scrolling graphs to libdazzle
These are the graphs we use in Builder. They are fairly strait forward,
although some slight of hand to try to keep CPU usage low. Some more we
can do. Although I suspect in a new wayland world we need to switch to
doing shaders as this is more expensive than XCopyArea() on X11.