theming: start unifying how we load theme data for libdazzle
This also switches back to "themes" instead of "css" for the theming
directory. I find that a bit easier to remember.

This moves a bunch of the theming stuff into a data directory which will
hopefully be easy to get contributors to add new theming support. One of
the main reasons Builder looks so poor on non-Adwaita is that we need to
encourage themes to help us out here, given all the custom widgets.

Although, maybe with dazzle broken out of Builder, it will be easier for
them to just support it from their theming files.