prefs: add preferences implementation
This is based on Builder's dynamic preferences engine. What Builder
provides on top of this is the Addin/Interface control to load plugins
which create these elements automatically (as well as revoke them).

We will stick to keeping that part downstream as it is dependent on the
design of the application (even if we are opinionated about using libpeas).