listbox: fix caching listbox implementation
This was not working as expected. This changes how we deal with
destruction by requiring that you subclass the list box row too.

When gtk_widget_destroy() is called, g_object_run_dispose() gets called.
We hook GObjectClass.dispose to check with the parent if we should first
hook into the cache.