fuzzy: pass last match offset when resolving and scoring
We want to use the last_offset to aide in scoring the document. So add
the plumbing to make this possible. We have some GSlice malloc overhead
on 32-bit systems, but on 64-bit it is all just pointer stashing for some
reduced overhead (same as what it was before).

Given that we aren't really targeting small 32-bit systems with this, I
think it's fine for now. If it becomes an issue, we can use a flywheel
array or something to reduce memory allocation overhead (along with a
32-bit int index into the array).