fuzzy: tweak scoring algorithm
We will use the last index of the match instead of strlen() to calculate
the score. This will naturally prefer keywords where the match is early
on to those where the match is later.

It also happens to reduce the strlen() cost (although that would be
trivialized by the malloc induced on 32-bit systems for now).