application: use consistent "resources://" prefix
Previously, we had mixed use of filesystem API and resource API. Instead,
this changes things to use resource:// explicitly as prefixes so that we
can also load files from disk. That might be necessary with plugins that
do not support GResources or for which GResources is not convenient.

This changes the function name of the public API entrypoint in
DzlApplication to make things a bit more apparent that it has changed
during this development cycle.