multi-paned: expand smaller children first
When we have multiple children who expand in the orientation
direction, we want to expand the smaller children first. This
gives us the effect of homogeneity as the allocation size grows.

I'm pretty sure we always want this. If for some reason not, we
can add a property to change the behavior. But I suspect that is
unlikely.