gtk: add helper to proxy GActionGroups from one widget to another
This can be useful when you have a situation with a HeaderBar and the
Content Area. You want the header bar to activate actions in the content
area, but they are not in a direct parent/child relationship. Therefore
the action muxer in Gtk cannot activate those actions.

This gets around that issue by looking at all the action groups associated
with the `from_widget` and ensuring them that they are copied to `widget`.

You can clear the old entry with a NULL `from_widget`.

Use a muxer key so that you can override previously muxed groups.