util: add liststore insertion with binary search
This will binary search the contents of @store to find the
location for a new row. @compare_func is used to compare the
content of a given column to @key.

When the appropriate location has been found, a new row is
inserted using gtk_list_store_insert().

The caller should use gtk_list_store_set() after calling this
function to populate the contents of the row.