list-store: directly access GSequence from GtkListStore
With knowledge of the GtkListStore implementation, we can
directly access the GSequenceIter in GtkTreeIter.user_data.
Doing so allows us to avoid traversing the GSequence treap
multiple times while inserting a value.

This is not guaranteed to do inserts exactly as a binary
search, as the GSequence approximates the mid-point based
on the sequence hierarchy.