properties-group: add flags for tweaking properties
This only has one tweak currently, which is to allow us to treat boolean
actions as requiring a parameter that is the new state. This allows us
to work around issues in Gtk where we want radio-button like features
but with just true/false.

Although, one can foresee features like 'invert-boolean' in the future.