application-window: start on DzlApplicationWindow
This is a base-window for applications to use that currently has
a single helper, a basic (un)fullscreen implementation. This
makes it easier on applications so they don't have to duplicate
this rather tricky logic.

One possible way to improve this would be to use a GdkWindow
that is input only with 5x high at the top of the screen. When
the revealer is expanded, increase the input-only window to
match the size. Then just watch for enter/leave notify.

As you can imagine, that is even more complex, so we'll avoid
doing that for now.