menu-button: add menu button helper
This is meant to be a base menu button class that we can use
for a few different situations in Builder. We want to be able
to replace the perspectives icon, the layoutstack icon (with
different theming), the run menu, and the gear menu with this.

It should handle the various styling we need as well as custom
GMenu information we support. It also needs to auto-update the
accelerator when that changes from keytheme changes.

This handles some of that, but continues to need iteration.