menu-manager: improved position solver
Previously we just had a very rudimentary solver for before-
after which couldn't deal with cases where peers were added
and new solutions needed to be found.

This improves on that with a 2-pass solver which attempts to
relocate items when new facts are found about the desired
positioning.

It's not perfect, but it's vastly better and easier to think
about.