shortcuts: add global activation phases
This adds an activation phase for global keybindings. You can have them as
either CAPTURE or BUBBLE phases. If CAPTURE, they run before the normal
capture phase of the hierarchy. If BUBBLE, then run after the BUBBLE phase
of the hierarchy.

There are still some things I wish we could do here, but getting all these
pieces to work well together has been quite tricky. Especially as we have
contexts, chords, and integration with the previous GtkBindingSets.

DzlShortcutPhase has been made a flags type, so we can bitwise-OR the
GLOBAL bit with the capture/bubble phases.

This breaks the ABI of the DzlShortcutEntry while we can still do that
without too much disruption. (Given that nothing is using this except
our branch of Builder).

I think we should add a property/attribute to the XML theme files that
allow us to have a phase specified for global keybindings. That is
currently a missing piece.

We might, someday, want to get really tricky with things and allow global
activation based on the current context. That is going to make things
much more tricky and I'm just not up to it right now.