shortcuts: take control of event delivery
Previously we were relying on binding sets and the event would
be delivered after the bubble phase if no binding entry matched.
That is not the behavior we want. We want actual event delivery
during the dispatch phase otherwise both capture/bubble can
override what the widget wants.

Additionally, this fixes an issue in Builder because it uses an
exotic extra fake widget to do CSS matching for event delivery.

Taking control of event delivery here allows that to continue
working by letting it dispatch the event to it's fake widget.