prefs: track destruction state of pref widgets
We want to be sure we don't call gtk_widget_destroy() on
something that is already in the destruction process. Also,
we want to ensure that we aren't holding onto pointers to
widgets after they've been destroyed.

This avoids that by using an indirection structure that allows
us to zero out the widget field on GtkWidget::destroy.