actions: add DZL_DEFINE_ACTION_GROUP()
This macro is useful to define GActionGroup's without having
to deal with all the nonsense of proxying to a backing
GSimpleActionGroup.

Additionally, it allows for non-annoying type signatures in
your action handlers.

Not a lot of docs yet, but that will land after I have some
more practical test cases in Builder.