treebuilder: add node-expanded and node-collapsed
This allows for builders to track the expand/collapse of nodes.
This can be useful to track external data.

We might want to do this for indirect nodes as well, but that
gets tricky in that we probably want to only do so when the
node is getting removed from rebuild-on-collapse.