recursive-monitor: add a recursive directory monitor
Since GFileMonitor uses a single inotify FD, we can use a bunch
of GFileMonitor while only using a single FD and not risk
hitting the fd ulimit. We still have a max watch limit with
inotify, but we're in better shape there as the max value is
much larger. (We probably race against tracker there, but
we'll see).