tree-node: add child helpers
This adds a dzl_tree_node_n_children() helper to get the number of children
for a node. It must have already been built to query this value.

It also adds dzl_tree_node_nth_child() which will get the @nth child of a
node and also requires the node have been built (although not the children
of the node).