file-monitor: improve ignored directory checks
If we take the cost of having O(h) directory FDs open (where `h'
is the height of the directory path from the root), then we can
reduce the number of ignored directory checks we must make. We
only need to walk past all the files with the ignored directory
as a path prefix.

https://bugzilla.gnome.org/show_bug.cgi?id=791400