tree: allow builders to hook into cell data func
This allows hooking into the cell data function so that nodes
do not need to contain static information. This can reduce lots
of allocations necessary to do custom text styling.

Currently, we only use this for the text cell func, but we can
probably add the pixbuf cell in the future. The signature
reflects that by using GtkCellRenderer and not the leaf type.