graph: avoid tick callbacks when were visible but obscured
There are plenty of cases where we can't really tell of we're visible and
obscured or actually visible to the user.

A heuristic we can use is to see how many frames we've had in tick_cb but
without a draw having landed. In that case, just disable the tick_cb until
we get to a real draw request (at which point it will start back up).