transient-grab: be safer when releasing a grab
We need to ensure that the grab does not disappear out from under us. We
also want to make sure that we have ownership of the widgets during the
processing of the events.

This adds a new reference to self during the operation, and references all
the potential widgets before doing notifications.