list-model-filter: add DzlListModelFilter
This is a filter for GListModel that is useful when using GtkListBox. It
allows for some basic filtering of the result sets. It does not, however
scale to gigantic lists, but could be made to if that was really an issue.

This has been used in Sysprof for a while now, and it seems generally
useful beyond that project.