suggestions: add rudimentary grab support
This moves the grab to the popover and then dispatches events back to
the entry. Doing so is a bit more like entry completion which could improve
our chances of avoiding xdg-popup focus weirdness.

Also, it ensures you can't press on a button from the main window while
the popover is visible, which is more inline with what the user would
expect given other gtk widgets. Clicking outside the drop down area will
cause the grab to break, and the popover to be dismissed.