widgets: do not reset child position on DzlMultiPaned::add
If we reset the positions each time a child is added, we don't allow
GtkBuilder .ui files to be able to set the position child property. That
is problematic since it requires apps to write code to set the
drag positions.

It doesn't seem to create much of an issue in practice to drop this.