backports: add GtkEventControllerMotion backport for pre-3.24
This is useful from a standpoint of being able to run on 3.22 such as
availiable on RHEL 8 and Debian stable.

The motion controller is pretty simple in nature. We just need to be able
to copy over the size of some private structs so that callback positions
line up properly.