search: find best match within corpus word
We might find a word that is a better match by starting with a character
later on in the word. However, if the previous scan moved us past the
character/pos tuple that would be necessary to match, we won't find it.

This allows us to rollback the position to a location where we can once
again match upcoming character for the current key-id. It's a small hit
(walking backwards) to greatly increase the likelyhood we match a
contiguous string.

This tweaks the scoring a bit so that we can detect a perfect match (all
characters one after another) without perturbing the score with the
length. Such is necessary so that longer words with a perfect match
do not score lower than shorter works with a partial match. However, the
in-between states there are going to fall into the previous trap (which
is probably fine for now).

Fixes #41