filemanager: create proxy synchronously
We can create the proxy without forcing the launch of the peer
process or waiting for signals/properties. Since this wont block
we meet the needs of !46 but without the extra main loop
processing. This is important because it allows
tests/test-file-manager to still work as all the necessary RPC
bytes will be written to the stream before exiting.