dzl-dock-widget: Don't set can-focus to True
Currently when moving the focus with the keyboard (tab, arrow keys), the
focus always moves in the Tab direction, without ever following what
focus move direction was requested.

This is because the DzlDockWidget tells it supports receiving the
keyboard focus, which is not the case: it is not e.g. a button or entry,
instead it's more a dock container than anything else.

So this commit fixes it by dropping the line that set the can-focus
property to True.